+++ src/components/nostr-short-text-note.ts
@@ -9,6 +9,7 @@ import {
   geohash2location,
   getEventGeohash,
 } from '../utils/geo.js'
+import './geo-direction-dialog.js'
 import './geo-direction.js'
 import './nostr-avatar.js'
 import './nostr-content.js'
@@ -58,11 +59,15 @@ export class NostrShortTextNote extends LitElement {
               <div>
                 ${this.origins.map(
                   o =>
-                    html`<geo-direction
+                    html`<geo-direction-dialog
                       .origin=${o.location}
-                      .dest=${dest}
-                      destGeohash=${ifDefined(geohash)}
-                    ></geo-direction> `,
+                      .destination=${geohash!}
+                      ><geo-direction
+                        .origin=${o.location}
+                        .dest=${dest}
+                        destGeohash=${ifDefined(geohash)}
+                      ></geo-direction
+                    ></geo-direction-dialog> `,
                 )}
                 ${typeof precision === 'number'
                   ? html`<span>&plusmn; ${formatDistance(precision)}</span>`